クラス名を整理
