+++ android/SoraAudioManager/SoraAudioManager/src/main/java/jp/shiguredo/sora/audiomanager/SoraAudioManager.java
@@ -12,41 +12,66 @@
 
 package jp.shiguredo.sora.audiomanager;
 
-import android.Manifest;
 import android.annotation.SuppressLint;
-import android.app.Activity;
+import android.content.BroadcastReceiver;
 import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.pm.PackageManager;
+import android.media.AudioAttributes;
+import android.media.AudioFocusRequest;
+import android.media.AudioManager;
 import android.os.Build;
 import android.util.Log;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+public class SoraAudioManager {
+    private static final String TAG = "SoraAudioManagerBase";
 
-public class SoraAudioManager extends SoraAudioManagerBase {
-    private static final String TAG = "SoraAudioManager";
-    private final SoraBluetoothManager bluetoothManager;
-    private Set<AudioDevice> audioDevices = new HashSet<>();
-    private boolean running;
-    private boolean savedIsSpeakerPhoneOn;
-    private AudioDevice selectedAudioDevice;
-    private AudioDevice lastConnectedAudioDevice;
-    private boolean willOffHandsfree;
+    public enum AudioDevice { SPEAKER_PHONE, WIRED_HEADSET, EARPIECE, BLUETOOTH, NONE }
+
+    protected final Context context;
+    protected final AudioManager audioManager;
+    protected final BroadcastReceiver wiredHeadsetReceiver;
+    protected final AudioDevice defaultAudioDevice;
+    protected boolean running;
+    private int savedAudioMode = AudioManager.MODE_INVALID;
+    private boolean savedIsMicrophoneMute;
+    protected boolean hasWiredHeadset;
+    protected boolean isSetHandsfree;
+    protected Object audioFocus;
+    protected OnChangeRouteObserver  onChangeRouteObserver;
 
     public interface OnChangeRouteObserver {
         // オーディオデバイスの変更を通知するコールバック
         void OnChangeRoute();
     }
 
-    static SoraAudioManager create(Context context) {
-        return new SoraAudioManager(context);
+    // 有線ヘッドセットの接続を通知するレシーバー
+    private class WiredHeadsetReceiver extends BroadcastReceiver {
+        private static final int STATE_UNPLUGGED = 0;
+        private static final int STATE_PLUGGED = 1;
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            int state = intent.getIntExtra("state", STATE_UNPLUGGED);
+            hasWiredHeadset = (state == STATE_PLUGGED);
+            updateAudioDeviceState();
+        }
     }
 
-    private SoraAudioManager(Context context) {
-        super(context);
-        bluetoothManager = SoraBluetoothManager.create(context, this, audioManager);
+    protected SoraAudioManager(Context context) {
+        SoraThreadUtils.checkIsOnMainThread();
+        this.context = context;
+        audioManager = (AudioManager) context.getSystemService(Context.AUDIO_SERVICE);
+        wiredHeadsetReceiver = new WiredHeadsetReceiver();
+        running = false;
+
+        // デフォルトのデバイスを設定する
+        // 受話用のスピーカーがある場合は受話用のスピーカーを使う
+        if (hasEarpiece()) {
+            defaultAudioDevice = AudioDevice.EARPIECE;
+        } else {
+            defaultAudioDevice = AudioDevice.SPEAKER_PHONE;
+        }
     }
 
     /*
@@ -57,196 +82,107 @@ public class SoraAudioManager extends SoraAudioManagerBase {
      * - モード
      * - マイクミュート
      */
-    @Override
-    public void start(OnChangeRouteObserver observer) {
-        Log.d(TAG, "start");
-        SoraThreadUtils.checkIsOnMainThread();
-        if (running) {
-            Log.e(TAG, "SoraAudioManager is already active");
-            return;
+    public void start(SoraAudioManagerLegacy.OnChangeRouteObserver  observer) {
+        savedAudioMode = audioManager.getMode();
+        savedIsMicrophoneMute = audioManager.isMicrophoneMute();
+
+        // オーディオフォーカスを取得する
+        // 前のオーディオフォーカス保持者に再生の一時停止を期待する
+        int result;
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            audioFocus = new AudioFocusRequest.Builder(AudioManager.AUDIOFOCUS_GAIN_TRANSIENT)
+                    .setAudioAttributes(
+                            new AudioAttributes.Builder()
+                                    .setUsage(AudioAttributes.USAGE_VOICE_COMMUNICATION)
+                                    .setContentType(AudioAttributes.CONTENT_TYPE_SPEECH)
+                                    .build()
+                    ).build();
+            result = audioManager.requestAudioFocus((AudioFocusRequest) audioFocus);
+        } else {
+            audioFocus = (AudioManager.OnAudioFocusChangeListener) focusChange -> {
+                // 取ったからといってその後何かするわけではないのでログだけ出す
+                Log.d(TAG, "onAudioFocusChange: " + focusChange);
+            };
+            result = audioManager.requestAudioFocus((AudioManager.OnAudioFocusChangeListener) audioFocus,
+                    AudioManager.STREAM_VOICE_CALL, AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
         }
-        running = true;
+        if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
+            Log.d(TAG, "Audio focus request granted for VOICE_CALL streams");
+        } else {
+            Log.e(TAG, "Audio focus request failed");
+        }
 
-        // コールバックを設定する
-        onChangeRouteObserver = observer;
+        // VoIP 向けのモードに切り替え
+        audioManager.setMode(AudioManager.MODE_IN_COMMUNICATION);
 
-        // 停止時に設定に戻すため設定を保存する
-        savedIsSpeakerPhoneOn = audioManager.isSpeakerphoneOn();
-        super.start(observer);
+        // マイクのミュートは解除する
+        setMicrophoneMute(false);
 
         // 初期化を行う
-        selectedAudioDevice = AudioDevice.NONE;
-        audioDevices.clear();
+        isSetHandsfree = false;
+    }
 
-        // bluetooth の制御を開始する
-        bluetoothManager.start();
-
-        // 初期化を行った状態でデバイスの設定を行う
-        // TODO(tnoho) この時点で初期値を指定しておきたいニーズが存在する可能性
-        updateAudioDeviceState();
-
-        registerWiredHeadsetReceiver();
+    // 有線ヘッドセットの接続を通知するレシーバーを登録する
+    protected void registerWiredHeadsetReceiver() {
+        context.registerReceiver(
+                wiredHeadsetReceiver,
+                new IntentFilter(Intent.ACTION_HEADSET_PLUG));
     }
 
     // オーディオの制御を終了する
-    @Override
     @SuppressLint("WrongConstant")
     public void stop() {
-        Log.d(TAG, "stop");
-        SoraThreadUtils.checkIsOnMainThread();
-        if (!running) {
-            Log.e(TAG, "Trying to stop SoraAudioManager in not running");
-            return;
-        }
-        running = false;
-
-        // 有線ヘッドセットの接続を通知するレシーバーを解除
-        unregisterWiredHeadsetReceiver();
-
-        // bluetooth の制御を終了する
-        bluetoothManager.stop();
-
         // 開始時に保存していた設定に戻す
-        setSpeakerphoneOn(savedIsSpeakerPhoneOn);
-        super.stop();
+        setMicrophoneMute(savedIsMicrophoneMute);
+        audioManager.setMode(savedAudioMode);
+
+        // オーディオフォーカスを放棄する
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            audioManager.abandonAudioFocusRequest((AudioFocusRequest) audioFocus);
+        } else {
+            audioManager.abandonAudioFocus((AudioManager.OnAudioFocusChangeListener) audioFocus);
+        }
+        audioFocus = null;
+
+        // コールバックを破棄する
+        onChangeRouteObserver = null;
     }
+
+    // 有線ヘッドセットの接続を通知するレシーバーを解除
+    protected void unregisterWiredHeadsetReceiver() {
+        context.unregisterReceiver(wiredHeadsetReceiver);
+    }
+
 
     // ハンズフリーかを確認する
     public boolean isHandsfree() {
-        // audioManager.isSpeakerphoneOn() は意図しない結果を返すのでこうした
-        return selectedAudioDevice == AudioDevice.SPEAKER_PHONE;
+        return false;
     }
 
-    // ハンズフリーの設定を変更する
-    private void setSpeakerphoneOn(boolean on) {
-        boolean wasOn = audioManager.isSpeakerphoneOn();
-        if (wasOn == on) {
+    // ハンズフリーに設定する
+    public void setHandsfree(boolean on) {
+        SoraThreadUtils.checkIsOnMainThread();
+        if (isSetHandsfree == on) {
             return;
         }
-        audioManager.setSpeakerphoneOn(on);
+        isSetHandsfree = on;
+        updateAudioDeviceState();
+    }
+
+    // マイクミュートの設定を変更する
+    private void setMicrophoneMute(boolean on) {
+        boolean wasMuted = audioManager.isMicrophoneMute();
+        if (wasMuted == on) {
+            return;
+        }
+        audioManager.setMicrophoneMute(on);
+    }
+
+    // 電話用スピーカーがデバイスにあるかを確認する
+    protected boolean hasEarpiece() {
+        return context.getPackageManager().hasSystemFeature(PackageManager.FEATURE_TELEPHONY);
     }
 
     // 状態に基づいてデバイスを選択する
-    @Override
-    public void updateAudioDeviceState() {
-        SoraThreadUtils.checkIsOnMainThread();
-        if (!running) {
-            return;
-        }
-
-        if (bluetoothManager.getState() == SoraBluetoothManager.State.HEADSET_AVAILABLE
-                || bluetoothManager.getState() == SoraBluetoothManager.State.HEADSET_UNAVAILABLE
-                || bluetoothManager.getState() == SoraBluetoothManager.State.SCO_DISCONNECTING) {
-            // Bluetooth のデバイスを更新する
-            bluetoothManager.updateDevice();
-        }
-
-        // 存在するオーディオデバイスのリストを生成する
-        Set<AudioDevice> newAudioDevices = new HashSet<>();
-
-        if (bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTED
-                || bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTING
-                || bluetoothManager.getState() == SoraBluetoothManager.State.HEADSET_AVAILABLE) {
-            // Bluetooth デバイスが存在する
-            newAudioDevices.add(AudioDevice.BLUETOOTH);
-        }
-
-        if (hasWiredHeadset) {
-            // 有線ヘッドセットが接続されている
-            newAudioDevices.add(AudioDevice.WIRED_HEADSET);
-        }
-
-        if (hasEarpiece()) {
-            // 受話用スピーカーが存在する
-            newAudioDevices.add(AudioDevice.EARPIECE);
-        }
-
-        // スピーカーは必ずある
-        newAudioDevices.add(AudioDevice.SPEAKER_PHONE);
-
-        // 利用可能なデバイス一覧に変更があったかを調べる
-        boolean audioDeviceSetUpdated = !audioDevices.equals(newAudioDevices);
-
-        if (audioDeviceSetUpdated) {
-            // 変更があった場合には新規で追加されたデバイスを抜き出す
-            List<AudioDevice> addedDevices = new ArrayList<>(newAudioDevices);
-            addedDevices.removeAll(audioDevices);
-
-            if (addedDevices.contains(AudioDevice.WIRED_HEADSET)) {
-                // 有線ヘッドセットが新たに接続された
-                lastConnectedAudioDevice = AudioDevice.WIRED_HEADSET;
-                // 新規デバイスが接続された場合はハンズフリーは解除する
-                isSetHandsfree = false;
-            } else if (addedDevices.contains(AudioDevice.BLUETOOTH)) {
-                // Bluetooth ヘッドセットが新たに接続された
-                lastConnectedAudioDevice = AudioDevice.BLUETOOTH;
-                if (isSetHandsfree) {
-                    // 新規デバイスが接続された場合はハンズフリーは解除する
-                    isSetHandsfree = false;
-                    // SCO_CONNECTED まではハンズフリーの解除は待つ
-                    willOffHandsfree = true;
-                }
-            }
-        }
-
-        // 利用可能なデバイス一覧を更新する
-        audioDevices = newAudioDevices;
-
-        // Bluetooth audio を開始する必要があるか
-        boolean needBluetoothAudioStart =
-                bluetoothManager.getState() == SoraBluetoothManager.State.HEADSET_AVAILABLE
-                        && !isSetHandsfree
-                        && (lastConnectedAudioDevice == AudioDevice.BLUETOOTH || !hasWiredHeadset);
-        Log.d(TAG, "--- updateAudioDeviceState: "
-                + "wired headset=" + hasWiredHeadset + ", "
-                + "BT state=" + bluetoothManager.getState() + ", "
-                + "needBluetoothAudioStart=" + needBluetoothAudioStart);
-
-        // Bluetooth audio を停止する必要があるか
-        boolean needBluetoothAudioStop =
-                (bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTED
-                        || bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTING)
-                        && (isSetHandsfree
-                        || (lastConnectedAudioDevice == AudioDevice.WIRED_HEADSET && hasWiredHeadset));
-
-        if (needBluetoothAudioStop) {
-            bluetoothManager.stopScoAudio();
-            bluetoothManager.updateDevice();
-        }
-
-        if (needBluetoothAudioStart && !needBluetoothAudioStop) {
-            // Bluetooth SCO audio を介しする
-            if (!bluetoothManager.startScoAudio()) {
-                // 失敗した場合はリストから BLUETOOTH を削除する
-                audioDevices.remove(AudioDevice.BLUETOOTH);
-                willOffHandsfree = false;
-            }
-        }
-
-        // ハンズフリーの解除を待つのは SCO_CONNECTING の間だけ
-        willOffHandsfree = willOffHandsfree
-                && bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTING;
-
-        final AudioDevice newAudioDevice;
-        if (bluetoothManager.getState() == SoraBluetoothManager.State.SCO_CONNECTED) {
-            newAudioDevice = AudioDevice.BLUETOOTH;
-            willOffHandsfree = false;
-        } else if (!isSetHandsfree && hasWiredHeadset) {
-            newAudioDevice = AudioDevice.WIRED_HEADSET;
-        } else if (isSetHandsfree || willOffHandsfree) {
-            newAudioDevice = AudioDevice.SPEAKER_PHONE;
-        } else {
-            newAudioDevice = defaultAudioDevice;
-        }
-        if (newAudioDevice != selectedAudioDevice) {
-            Log.d(TAG, "New device status: "
-                    + "available=" + audioDevices + ", "
-                    + "selected=" + newAudioDevice);
-            setSpeakerphoneOn(newAudioDevice == AudioDevice.SPEAKER_PHONE);
-            selectedAudioDevice = newAudioDevice;
-            if (onChangeRouteObserver != null) {
-                onChangeRouteObserver.OnChangeRoute();
-            }
-        }
-    }
+    public void updateAudioDeviceState() {}
 }
